Add: task 2

+++ 0x0C-more_malloc_free/2-calloc.c
@@ -11,7 +11,7 @@
 
 void *_calloc(unsigned int nmemb, unsigned int size)
 {
-	int i;
+	unsigned int i;
 	char *p;
 
 	if (nmemb == 0 || size == 0)
@@ -21,7 +21,8 @@ void *_calloc(unsigned int nmemb, unsigned int size)
 	if (p == NULL)
 		return (NULL);
 
-	for (i = 0; i < (int)nmemb; i++)
+	for (i = 0; i < nmemb; i++)
 		*(p + i) = 0;
+
 	return ((void *)p);
 }
